pyproject.toml
[build-system]
requires = ["maturin>=1.13.1"]
build-backend = "maturin"

[project]
name = "polars-lance"
description = "Polars plugin for reading and writing Lance datasets"
version = "0.2.0"
authors = [{ name = "Jorrit Sandbrink" }]
keywords = ["polars", "lance", "arrow"]
requires-python = ">=3.10"
readme = "README.md"
license = "MIT"
license-files = ["LICENSE"]
dependencies = [
  "polars>=1.0.0",
]
classifiers = [
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3.13",
  "Programming Language :: Python :: 3.14",
  "Programming Language :: Rust",
]
urls = { Repository = "https://github.com/jorritsandbrink/polars-lance" }

[dependency-groups]
dev = [
  "maturin>=1.13.1",
  "pylance>=4.0.1",
  "pyarrow>=14.0.0",
  "pytest>=8.0.0",
  "mypy>=1.20.2",
  "ruff>=0.15.12",
  "testcontainers[minio]>=4.14.2",
  "azure-storage-blob>=12.29.0",
]

[tool.pytest.ini_options]
markers = ["needs_docker: marks tests that require a running Docker daemon"]
filterwarnings = [
  "ignore:The @wait_container_is_ready decorator is deprecated:DeprecationWarning:testcontainers",
]

[tool.maturin]
python-source = "python"
module-name = "polars_lance._polars_lance"
features = ["pyo3/extension-module", "pyo3"]

[tool.mypy]
files = ["python/polars_lance", "tests"]
disallow_untyped_defs = true

[[tool.mypy.overrides]]
module = "testcontainers.*"
ignore_missing_imports = true

[tool.ruff]
extend-exclude = ["target"]
